python/requirements.txt
torch==1.10.1
numpy==1.21.0
sympy==1.9
opencv-python==4.5.5.62
logzero==1.7.0
scikit-image==0.19.1
scipy==1.7.3
Pillow==9.0.0
pydicom==2.2.2
segmentation-models-pytorch==0.2.1